Refactors, printing regs, several instructions
Refactored some of the code to keep it DRY. Added send_areg and
send_dreg for printing registers. A few more instructions supported such
as LEA and NEG. As soon as I write something to differentiate OPMODEs,
that will knock out like 8 more instructions and the reg printing now
will be quite trivial.